refactor: test_util
