+++ src/query/src/lib.rs
@@ -37,6 +37,7 @@ mod range_select;
 pub mod region_query;
 pub mod sql;
 pub mod stats;
+pub(crate) mod test_util;
 pub(crate) mod window_sort;
 
 #[cfg(test)]
